Delta2 (#26)
* deconstruct sigreader returns

* adjust a couple things in blocksig.. vectorize equals?

* combine blockMatcher and lazyBlockSig

* start skeleton for deltaWriter2

* figuring out the loop structure

* figure out some inner loop details

* more inner loop pieces, start commenting old stuff

* part of slidingBlock internals

* finish first draft of slidingBlock internals

* figuring out the deltaWriter outer loop

* more of outer deltaWriter loop

* finish most of outerloop

* move rolling hash updates into slidingBlock

* cleanup: remove commented code, rolling hash interface changes

* more inner loop implementation

* unit tests passing

* add some finish asserts

* less offset math in outer loop

* fix lazy strong hash calculation

* fix pending literal calculation

* implement max literal length check

+++ rsync.delta/models/CopyCommand.cs
@@ -11,12 +11,14 @@ namespace Rsync.Delta.Models
         private readonly CommandArg _start;
         private readonly CommandArg _length;
 
-        public LongRange Range => new LongRange(_start.Value, _length.Value);
+        public LongRange Range => new LongRange(
+            (long)_start.Value, 
+            (long)_length.Value);
 
         public CopyCommand(LongRange range)
         {
-            _start = new CommandArg(range.Start);
-            _length = new CommandArg(range.Length);
+            _start = new CommandArg((ulong)range.Start);
+            _length = new CommandArg((ulong)range.Length);
         }
 
         private CopyCommand(CommandArg startArg, CommandArg lengthArg)
